requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
channels
django-cors-headers
django-environ
daphne
channels_redis
pillow